pyproject.toml
[project]
name = "decode-base64-api"
version = "0.1.0"
description = ""
authors = [
    {name = "telesinfo",email = "telesinfo@telesinfo.com.br"}
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.12"
uvicorn = "^0.27.0"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"

[tool.poetry]
packages = [{include = "decode_base64_api", from = "src"}]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
blue = "^0.9.1"
isort = "^6.0.1"
taskipy = "^1.14.1"


[tool.poetry.group.doc.dependencies]
mkdocs-material = "^9.6.12"
mkdocstrings = "^0.29.1"
mkdocstrings-python = "^1.16.10"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
